% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{splot}
\alias{splot}
\title{Split Plot}
\usage{
splot(y, x = NULL, by = NULL, between = NULL, cov = NULL, type = "",
  split = "median", data = NULL, su = NULL, levels = list(),
  error = "standard", error.color = "#585858", lim = 9, lines = TRUE,
  mv.scale = "none", mv.as.x = FALSE, save = FALSE, format = cairo_pdf,
  dims = dev.size(), file.name = "splot", colors = NULL, myl = NULL,
  mxl = NULL, autori = TRUE, xlas = 0, ylas = 1, lwd = 2, pch = 20,
  bw = "nrd0", adj = 2, lpos = "auto", lvn = TRUE, title = TRUE,
  labx = TRUE, laby = TRUE, lty = TRUE, lhz = FALSE, sub = TRUE,
  ndisp = TRUE, leg = TRUE, note = TRUE, sud = TRUE, labels = TRUE,
  labels.filter = "_|\\\\.", points = TRUE, points.first = TRUE,
  byx = TRUE, drop = TRUE, model = FALSE, mar = "auto", add = NULL,
  ...)
}
\arguments{
\item{y}{required. Primary variable, to be shown on the y axis unless \code{x} is not specified. Can be an object,
name of a column in data, or a formula (see note). Multiple variables can also be included in a matrix-like object.}

\item{x}{secondary variable, to be shown in on the x axis. If not specified, \code{type} will be set to \code{'density'}.
If \code{x} is a factor or vector of characters, or has fewer than \code{lim} levels when treated as a factor, \code{type}
will be set to \code{"line"} unless specified.}

\item{by}{the 'splitting' variable within each plot, by which the plotted values of \code{x} and \code{y} will be grouped.}

\item{between}{a single object or name, or two in a vector (e.g., \code{c(b1, b2)}), the levels of which will determine the
number of plot windows to be shown at once (the cells in a matrix of plots; levels of the first variable as rows, and
levels of the second as columns).}

\item{cov}{additional variables used for adjustment. Bar and line plots include all \code{cov} variables in their regression models (via
\code{\link[stats]{lm}}, e.g., \code{lm(y ~ 0 + x + cov1 + cov2)}) as covariates. Scatter plots with lines include all \code{cov} variables
in the regression model to adjust the prediction line (e.g., \code{lm(y ~ x + x^2)}).}

\item{type}{determines the type of plot to make, between \code{"bar"}, \code{"line"}, \code{"density"}, or \code{"scatter"}. If
\code{"density"}, \code{x} is ignored. Anything including the first letter of each is accepted (e.g., \code{type='l'}).}

\item{split}{how to split any continuous variables (those with more than \code{lim} levels as factors). Default is \code{"median"},
with \code{"mean"}, \code{"standard deviation"}, and \code{"quantile"} as options.}

\item{data}{a \code{data.frame} to pull variables from. If variables aren't found in \code{data}, they will be looked for in the environment.}

\item{su}{a subset to all variables, applied after they are all retrieved from \code{data} or the environment.}

\item{levels}{a list with entries corresponding to variable names, used to rename and/or reorder factor levels. To reorder a factor, enter a
vector of either numbers or existing level names in the new order (e.g., \code{levels=list(var=c(3,2,1))}). To rename levels of a factor,
enter a character vector the same length as the number of levels. To rename and reorder, enter a list, with names as the first entry, and
order as the second entry (e.g., \code{levels=list(var=list(c('a','b','c'),c(3,2,1)))}). This happens after variables are split, so
names and orders should correspond to the new split levels of split variables. For example, if a continuous variable is median split, it now
has two levels ('Under Median' and 'Over Median'), which are the levels reordering or renaming would apply to.}

\item{error}{determines whether and which type of error bars to show in bar or line plots. If \code{FALSE}, no error bars will be shown.
Otherwise, the default is \code{"standard error"}, with \code{"confidence intervals"} as an option.}

\item{error.color}{color of the error bars. Default is \code{'#585858'}.}

\item{lim}{numeric. Checked against the number of factor levels of each variable. Used to decide which variables should be split, which colors
to use, and when to turn off the legend. Default is \code{9}. If set over \code{20}, \code{lim} is treated as infinite (set to \code{Inf}).}

\item{lines}{logical or a string specifying the type of lines to be drawn in scatter plots. By default (and whenever \code{cov} is not missing),
a prediction line is fitted with \code{\link[stats]{lm}}. For (potentially) bendy lines, \code{'loess'} (matching \code{'^lo|^p|^cu'}) will use
\code{\link[stats]{loess}}, and \code{'spline'} (\code{'^sm|^sp|^in'}) will use \code{\link[stats]{smooth.spline}}. \code{'connected'}
(\code{'^e|^co|^d'}) will draw lines connecting all points, and \code{FALSE} will not draw any lines.}

\item{mv.scale}{determines whether to center and scale multiple \code{y} variables. Does not center or scale by default. Anything other than
\code{'none'} will mean center each numeric \code{y} variable. Anything matching \code{'^t|z|sc'} will also scale.}

\item{mv.as.x}{logical: if \code{TRUE}, variable names are displayed on the x axis, and \code{x} is treated as \code{by}.}

\item{save}{logical: if \code{TRUE}, an image of the plot is saved to the current working directory.}

\item{format}{the type of file to save plots as. default is \code{\link[grDevices]{cairo_pdf}}. See \code{\link[grDevices]{Devices}} for options.}

\item{dims}{a vector of 2 values (\code{c(width, height)}) specifying the dimensions of a plot to save in inches or pixels depending on
\code{format}. Defaults to the dimensions of the plot window.}

\item{file.name}{a string with the name of the file to be save (excluding the extension, as this is added depending on \code{format}).}

\item{colors}{sets a color theme or manually specifies colors. Default theme is \code{"pastel"}, with \code{"dark"} and \code{"bright"} as
options. If set to \code{"grey"}, or if \code{by} has more than 9 levels, a grey scale is calculated using \code{\link[grDevices]{grey}}.
See the \code{col} parameter in \code{\link[graphics]{par}} for acceptable manual inputs.}

\item{myl}{sets the range of the y axis (\code{ylim} of \code{\link[graphics]{plot}} or \code{\link[graphics]{barplot}}). If not specified,
this will be calculated from the data.}

\item{mxl}{sets the range of the x axis (\code{xlim} of \code{\link[graphics]{plot}}). If not specified, this will be calculated from the data.}

\item{autori}{logical: if \code{FALSE}, the origin of plotted bars will be set to 0. Otherwise, bars are adjusted such that they extend to the
bottom of the y axis.}

\item{xlas}{sets the orientation of the x-axis labels. See \code{\link[graphics]{par}}.}

\item{ylas}{sets the orientation of the y-axis labels. See \code{\link[graphics]{par}}.}

\item{lwd}{sets the width of lines in scatter plots. Default is 2. See \code{\link[graphics]{par}}.}

\item{pch}{sets the type of dot to use when plotting points in a scatter plot. Default is 20. See \code{\link[graphics]{par}}.}

\item{bw}{sets the smoothing bandwidth when plotting densities. Default is \code{'nrd0'}. See \code{\link[stats]{density}}.}

\item{adj}{adjusts the smoothing of densities (\code{adj * bw}). See \code{\link[stats]{density}}.}

\item{lpos}{set the position of the legend. By default, it will try to choose between the top left and right areas. See \code{\link{legend}}.}

\item{lvn}{legend variable name. Logical: if \code{FALSE}, the names of by and between variables will not be shown before their level (e.g.,
for a sex variable with a "female" level, "sex: female" would become "female" in the legend or above each plot window).}

\item{title}{logical or a character: if \code{FALSE}, the main title is turned off. If a character, this will be shown as the main title.}

\item{labx}{logical or a character: if \code{FALSE}, the label on the x axis is turned off. If a character, this will be shown as the
x axis label.}

\item{laby}{logical or a character: if \code{FALSE}, the label on the y axis is turned off. If a character, this will be shown as the
y axis label.}

\item{lty}{logical or a vector: if \code{FALSE}, lines are always solid. If a vector, changes line type based on each value. Otherwise loops
through available line types, see \code{\link[graphics]{par}}.}

\item{lhz}{logical: if \code{FALSE}, the legend grows horizontally.}

\item{sub}{logical: if \code{FALSE}, the small title above each plot showing \code{between} levels is turned off.}

\item{ndisp}{logical: if \code{FALSE}, n per level is no longer displayed in the subheadings.}

\item{leg}{logical: if \code{FALSE}, the legend is turned off.}

\item{note}{logical: if \code{FALSE}, the note at the bottom about splits and/or error bars is turned off.}

\item{sud}{logical: if \code{FALSE}, the heading for subset and covariates/line adjustments is turned off.}

\item{labels}{logical: if \code{FALSE}, sets all settable text surrounding the plot to \code{FALSE} (just so you don't have to set all
of them if you want a clean frame).}

\item{labels.filter}{a regular expression string to be replaced in label texts with a blank space. Default is \code{'_|\\\\.'}; underscores and
periods appearing in the text of labels are replace with blank spaces. Quotation marks are removed in addition to the \code{labels.filter}
string. Set to \code{FALSE} to prevent all filtering.}

\item{points}{logical: if \code{FALSE}, the points in a scatter plot are no longer drawn.}

\item{points.first}{logical: if \code{FALSE}, points are plotted after lines are drawn in a scatter plot, placing lines behind points. This does
not apply to points or lines added in \code{add}, as that is always evaluated after the main points and lines are drawn.}

\item{byx}{logical: if \code{TRUE} (default) and \code{by} is specified, regressions for bar or line plots compare levels of \code{by} for each
level of \code{x}. This makes for more intuitive error bars when comparing levels of \code{by} within a level of \code{x}.}

\item{drop}{logical: if \code{FALSE}, \code{x} levels with no data are still displayed in bar or line plots.}

\item{model}{logical: if \code{TRUE}, the summary of an interaction model will be printed.}

\item{mar}{sets the margins of each plot window. Partially set automatically if not specified: \code{c(if(labx)2.5 else 0,if(laby)3 else 2,1,0)}.
If \code{xlas} is not specified, or is greater than 2, and x-axis labels are overly long in bar or line plots,
\code{mar[1]} is set by the x-axis text length (\code{strwidth(max(colnames(m)))*ifelse(labx,5.5,4.8)}). See \code{\link[graphics]{par}}.}

\item{add}{evaluated within the function. Useful for adding things like lines to a plot while the parameters are still those set by the
function (e.g., \code{add=\{lines(1:10)\}}).}

\item{...}{passes additional arguments to \code{\link[graphics]{plot}} or \code{\link[graphics]{barplot}}.}
}
\description{
A plotting function aimed at automating some common visualization tasks in order to ease data exploration.
}
\note{
When \code{y} is a formula, other variables will be pulled from it:

\code{y ~ x * by * between[1] * between[2] + cov[1] + cov[2] + cov[n]}

If \code{y} has multiple variables, \code{by} is used to identify the variable (it becomes a factor with variable names as levels),
so anything entered as \code{by} is treated as \code{between[1]}, \code{between[1]} is moved to \code{between[2]}, and \code{between[2]}
is discarded with a message.
}
\examples{
#simulating data
n=2000
dat=data.frame(sapply(c('by','bet1','bet2'),function(c)sample(0:1,n,TRUE)))
dat$x=eval(quote(
  rnorm(n)+by*-.4+by*bet1*-.3+by*bet2*.3+bet1*bet2*.9-.8+rnorm(n,0,by)
),envir=dat)
dat$y=eval(quote(
  x*.2+by*.3+bet2*-.6+bet1*bet2*.8+x*by*bet1*-.5+x*by*bet1*bet2*-.5
  +rnorm(n,5)+rnorm(n,-1,.1*x^2)
),envir=dat)

#looking at the distribution of y between bets split by by
splot(y, by=by, between=c(bet1, bet2), data=dat)

#looking at quantile splits of y in y by x
splot(y~x*y, split='quantile', data=dat)

#looking at y by x between bets
splot(y~x, between=c(bet1, bet2), data=dat)

#sequentially adding levels of split
splot(y~x*by, data=dat)
splot(y~x*by*bet1, data=dat)
splot(y~x*by*bet1*bet2, data=dat)

#same as the last but entered differently
splot(y, x, by, c(bet1, bet2), data=dat)

#zooming in on one of the windows
splot(y~x*by, data=dat, su=bet1==1&bet2==0)

#compairing an adjusted lm prediction line with a loess line
#this could also be entered as y ~ poly(x,3)
splot(y~x+x^2+x^3, data=dat, su=bet1==1&bet2==0&by==1, add={
  lines(x[order(x)], loess(y~x)$fitted[order(x)], lty=2)
  legend('topright', c('lm', 'loess'), lty=c(1, 2), lwd=c(2, 1), bty='n')
})

#looking at different versions of x added to y
splot(cbind(
  Raw=y+x,
  Sine=y+sin(x),
  Cosine=y+cos(x),
  Tangent=y+tan(x)
)~x, data=dat, myl=c(-10,15), lines='loess', lpos='topright')

}
